Updated to lit3.
properties don't need to be defined as functions any longer, just an
object literal will do. Boolean attributes (fav) shouldn't be cast about
into different types if you can help it. Also, boolean attributes don't
need a string value like "true". They're either there or they're not.
I simplified the render method to not be as repetitive. You can use the
variable in a js expression to modify the single template output. If you
update "fav" on the tag, Lit will just rerender the bare minimum it
needs to as it understands the template parts.